requirements.txt
google-adk
httpx
mcp
uvicorn
fastapi
